Move pkg/cli -> internal/cli
